Many minor fixes. Still have problems with %g precision and subnormal numbers.

+++ drafts/libft/src/math/op/ft_floor.c
@@ -12,10 +12,12 @@
 
 #include "libft.h"
 
+//TODO remove for moulinette
+
 long		ft_floor(double lf)
 {
 	long	result;
-	short	exp;
+	int		exp;
 
 	exp = (((t_u64)lf << 1) >> 52) - 1023;
 	if (exp > 308)
